requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.25
mysql-connector-python==9.0.0
python-dotenv==1.0.1

